requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# pytorch-3d-r2n2/test.py: 6
# pytorch-3d-r2n2/train.py: 13
ConfigArgParse == 0.14.0

# pytorch-3d-r2n2/dataset.py: 7
Pillow == 9.3.0

# pytorch-3d-r2n2/binvox_rw.py: 65
# pytorch-3d-r2n2/dataset.py: 8
# pytorch-3d-r2n2/lib/test_minibatch.py: 4
# pytorch-3d-r2n2/lib/validate.py: 3
# pytorch-3d-r2n2/utils.py: 1
numpy == 1.22.0

# pytorch-3d-r2n2/train.py: 9
tensorboardX == 1.7

# pytorch-3d-r2n2/dataset.py: 4,6
# pytorch-3d-r2n2/lib/resume.py: 1
# pytorch-3d-r2n2/lib/test_minibatch.py: 1
# pytorch-3d-r2n2/lib/validate.py: 1
# pytorch-3d-r2n2/modules/conv_rnn.py: 1,2,3
# pytorch-3d-r2n2/network.py: 1
# pytorch-3d-r2n2/test.py: 3
# pytorch-3d-r2n2/train.py: 4,5,6,7
torch == 1.4.0

# pytorch-3d-r2n2/dataset.py: 5
# pytorch-3d-r2n2/test.py: 4
# pytorch-3d-r2n2/train.py: 8,10
torchvision == 0.5.0

EasyDict
pyyaml
sklearn
